requirements.txt
pytest==8.3.3
allure-pytest==2.13.5
selene>=2.0.0rc4
Appium-Python-Client==4.4.0
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1
requests==2.32.3